move player related stuff to player file
